% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna2prt.R
\docType{data}
\name{dna}
\alias{dna}
\title{Convert nucleotide sequence to amino acid sequence}
\format{An object of class \code{character} of length 1.}
\usage{
dna
}
\arguments{
\item{dna}{A character vector containing a single text string of DNA}
}
\value{
A character vector containing a single text string of amino acids
}
\description{
Convert nucleotide sequence to amino acid sequence
}
\keyword{datasets}
